fix(migration): heal_succeeded is int, not tinyint(1) — match schema
Migration 0003 declared heal_succeeded as tinyint(1) but the drizzle
schema at task-steps.ts uses `int('heal_succeeded')`. On any machine
that ran the migration against an empty DB and then `db:push`,
drizzle-kit proposes a data-loss truncate to widen the column type.

Harmless for fresh installs from HEAD (push converges the first
time), but painful for anyone who applied 0003 via migrate and then
tried push — you lose your task_steps rows unless you hand-apply a
MODIFY COLUMN first.

Change the ADD column DDL to `int DEFAULT 0 NOT NULL` so new
installs write the right type up front. Existing installs with
tinyint(1) should run:

  ALTER TABLE task_steps MODIFY COLUMN heal_succeeded int NOT NULL DEFAULT 0;

to converge without data loss (tinyint->int is a widening, safe).

+++ apps/orchestrator/drizzle/0003_heal_metrics.sql
@@ -1,5 +1,5 @@
 ALTER TABLE `task_steps` ADD `heal_attempts` int DEFAULT 0 NOT NULL;--> statement-breakpoint
-ALTER TABLE `task_steps` ADD `heal_succeeded` tinyint(1) DEFAULT 0 NOT NULL;--> statement-breakpoint
+ALTER TABLE `task_steps` ADD `heal_succeeded` int DEFAULT 0 NOT NULL;--> statement-breakpoint
 ALTER TABLE `task_steps` ADD `heal_elapsed_ms` int;--> statement-breakpoint
 ALTER TABLE `task_steps` ADD `heal_input_tokens` int;--> statement-breakpoint
 ALTER TABLE `task_steps` ADD `heal_output_tokens` int;--> statement-breakpoint
